chore: Apply minor changes to comply peer review
