+++ package.json
@@ -29,7 +29,6 @@
     "bcryptjs": "^2.4.3",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "express-http-context": "^1.2.4",
     "graphql": "^16.0.0",
     "jsonwebtoken": "^8.5.1",
     "nodemailer": "^6.7.0",
