+++ src/exceptions/auth/auth.codes.ts
@@ -2,3 +2,4 @@ export const USER_NOT_FOUND = '1001';
 export const INVALID_EMAIL = '1002';
 export const USER_COULD_NOT_BE_CREATED = '1003';
 export const USER_COULD_NOT_BE_AUTHENTICATED = '1004';
+export const MISSING_TOKEN = '1005';
