+++ src/graphql/models/Mutations.ts
@@ -13,6 +13,6 @@ export const Mutation = `
     updateProject(projectInput: UpdateProjectInput!): Project!
     createTaskStatus(taskStatusInput: TaskStatusInput!): TaskStatus!
     assignTaskStatus(taskId: String!, taskStatusId: String!): Task!
-    deleteUser(email: String!, password: String!): Message!
+    deleteUser(email: String!, password: String!): Void!
   }
 `;
