+++ src/interfaces.ts
@@ -17,10 +17,19 @@ export interface LoginArgs {
     password: string,
 }
 
-export interface DeleteUserArgs {
-  email: string,
-  password: string,
-}
+// An interface declaring no members is equivalent to its supertype.
+export type DeleteUserArgs = LoginArgs
+
+export interface UpdateUserByIdArgs {
+    id: string,
+    data: {
+      password?: string,
+      lastPasswordReset?: Date,
+      last_activity?: Date,
+      is_active: boolean,
+      name?: string
+    }
+  }
 
 export interface Sprint {
     id: string;
