+++ src/repositories/AuthenticationRepository/index.ts
@@ -2,3 +2,4 @@ export * from './createOneUser.repository';
 export * from './oneUserByEmail.repository';
 export * from './oneUserById.repository';
 export * from './deleteOneUserById.repository';
+export * from './updateOneUserById.repository';
