+++ src/server.ts
@@ -40,7 +40,7 @@ async function startServer() {
         },
     });
 
-    app.listen(PORT, () => console.log(`Server is running on port ${4000}`));
+    app.listen(PORT, () => console.log(`Server is running on port ${PORT}`));
 }
 
 startServer();
